Break up the code related to various algorithms and marginal tree iteration into different header files.
