+++ msprime_algo.hpp
@@ -16,245 +16,16 @@
 /// Author: Kevin Thornton
 /// Thanks: Jerome Kelleher
 
-#include <algorithm>
 #include <vector>
 #include <cstdint>
-#include <tuple>
-#include <limits>
-#include "indexed_edge.hpp"
+#include <type_traits>
+#include "marginal_tree.hpp"
+#include "iterate_marginal_trees.hpp"
 
 namespace fwdpp
 {
     namespace ts
     {
-        struct marginal_tree
-        {
-            std::vector<std::int32_t> parents, leaf_counts, left_sib,
-                right_sib, left_child, right_child, left_sample, right_sample,
-                next_sample, is_sample;
-            double left, right;
-            marginal_tree(std::int32_t nnodes,
-                          const std::vector<std::int32_t>& samples)
-                : parents(nnodes, -1), leaf_counts(nnodes, 0),
-                  left_sib(nnodes, -1), right_sib(nnodes, -1),
-                  left_child(nnodes, -1), right_child(nnodes, -1),
-                  left_sample(nnodes, -1), right_sample(nnodes, -1),
-                  next_sample(nnodes, -1), is_sample(nnodes, 0),
-                  left{ std::numeric_limits<double>::quiet_NaN() }, right{
-                      std::numeric_limits<double>::quiet_NaN()
-                  }
-            {
-                for (auto s : samples)
-                    {
-                        if (static_cast<std::size_t>(s) >= leaf_counts.size())
-                            {
-                                throw std::invalid_argument(
-                                    "sample index out of range");
-                            }
-                        leaf_counts[s] = 1;
-                        is_sample[s] = 1;
-                        left_sample[s] = right_sample[s] = s;
-                    }
-            }
-            marginal_tree(std::int32_t nnodes)
-                : parents(nnodes, -1), leaf_counts{}, left_sib(nnodes, -1),
-                  right_sib(nnodes, -1), left_child(nnodes, -1),
-                  right_child(nnodes, -1), left_sample(nnodes, -1),
-                  right_sample(nnodes, -1), next_sample(nnodes, -1),
-                  is_sample(nnodes, 0),
-                  left{ std::numeric_limits<double>::quiet_NaN() }, right{
-                      std::numeric_limits<double>::quiet_NaN()
-                  }
-            {
-            }
-        };
-
-        void
-        outgoing_leaf_counts(marginal_tree&, const std::int32_t,
-                             const std::int32_t, const std::false_type)
-        // TODO: internal namespace
-        {
-        }
-
-        inline void
-        incoming_leaf_counts(marginal_tree&, const std::int32_t,
-                             const std::int32_t, const std::false_type)
-        // TODO: internal namespace
-        {
-        }
-
-        inline void
-        outgoing_leaf_counts(marginal_tree& marginal,
-                             const std::int32_t parent,
-                             const std::int32_t child, const std::true_type)
-        // TODO: internal namespace
-        {
-            auto p = parent;
-            auto lc = marginal.leaf_counts[child];
-            while (p != -1)
-                {
-                    marginal.leaf_counts[p] -= lc;
-                    assert(marginal.leaf_counts[p] >= 0);
-                    p = marginal.parents[p];
-                }
-        }
-
-        inline void
-        incoming_leaf_counts(marginal_tree& marginal,
-                             const std::int32_t parent,
-                             const std::int32_t child, const std::true_type)
-        // TODO: internal namespace
-        {
-            auto p = parent;
-            auto lc = marginal.leaf_counts[child];
-            while (p != -1)
-                {
-                    marginal.leaf_counts[p] += lc;
-                    p = marginal.parents[p];
-                }
-        }
-
-        inline void
-        update_sample_list(marginal_tree& marginal, const std::int32_t parent,
-                           const std::true_type)
-        // TODO: internal namespace
-        // TODO: make this a standalone function rather than an updating policy
-        {
-            for (auto n = parent; n != -1; n = marginal.parents[n])
-                {
-                    if (marginal.is_sample[n] == 1)
-                        {
-                            marginal.left_sample[n] = marginal.right_sample[n];
-                        }
-                    else
-                        {
-                            marginal.left_sample[n] = marginal.right_sample[n]
-                                = -1;
-                        }
-                    for (auto v = marginal.left_child[n]; v != -1;
-                         v = marginal.right_sib[v])
-                        {
-                            if (marginal.left_sample[v] != -1)
-                                {
-                                    assert(marginal.right_sample[v] != -1);
-                                    if (marginal.left_sample[n] == -1)
-                                        {
-                                            marginal.left_sample[n]
-                                                = marginal.left_sample[v];
-                                            marginal.right_sample[n]
-                                                = marginal.right_sample[v];
-                                        }
-                                    else
-                                        {
-                                            marginal.next_sample
-                                                [marginal.right_sample[n]]
-                                                = marginal.left_sample[v];
-                                            marginal.right_sample[n]
-                                                = marginal.right_sample[v];
-                                        }
-                                }
-                        }
-                }
-        }
-
-        inline void
-        update_sample_list(marginal_tree&, const std::int32_t,
-                           const std::false_type)
-        // TODO: internal namespace
-        {
-        }
-
-        template <typename visitor, typename leaf_policy,
-                  typename sample_list_policy>
-        inline void
-        iterate_marginal_trees(marginal_tree& marginal,
-                               const indexed_edge_container& input_left,
-                               const indexed_edge_container& output_right,
-                               const double maxpos, visitor v,
-                               const leaf_policy lp,
-                               const sample_list_policy slp)
-        // TODO: internal namespace
-        {
-            auto j = input_left.begin(), jM = input_left.end(),
-                 k = output_right.begin(), kM = output_right.end();
-            double x = 0.0;
-            while (j < jM || x < maxpos)
-                {
-                    while (k < kM && k->pos == x) // T4
-                        {
-                            const auto p = k->parent;
-                            const auto c = k->child;
-                            const auto lsib = marginal.left_sib[c];
-                            const auto rsib = marginal.right_sib[c];
-                            if (lsib == -1)
-                                {
-                                    marginal.left_child[p] = rsib;
-                                }
-                            else
-                                {
-                                    marginal.right_sib[lsib] = rsib;
-                                }
-                            if (rsib == -1)
-                                {
-                                    marginal.right_child[p] = lsib;
-                                }
-                            else
-                                {
-                                    marginal.left_sib[rsib] = lsib;
-                                }
-                            marginal.parents[c] = -1;
-                            marginal.left_sib[c] = -1;
-                            marginal.right_sib[c] = -1;
-                            outgoing_leaf_counts(marginal, k->parent, k->child,
-                                                 lp);
-                            update_sample_list(marginal, k->parent, slp);
-                            ++k;
-                        }
-                    while (j < jM && j->pos == x) // Step T2
-                        {
-                            const auto p = j->parent;
-                            const auto c = j->child;
-                            const auto rchild = marginal.right_child[p];
-                            if (rchild == -1)
-                                {
-                                    marginal.left_child[p] = c;
-                                    marginal.left_sib[c] = -1;
-                                    marginal.right_sib[c] = -1;
-                                }
-                            else
-                                {
-                                    marginal.right_sib[rchild] = c;
-                                    marginal.left_sib[c] = rchild;
-                                    marginal.right_sib[c] = -1;
-                                }
-                            // The entry for the child refers to
-                            // the parent's location in the node table.
-                            marginal.parents[c] = j->parent;
-                            marginal.right_child[p] = c;
-                            incoming_leaf_counts(marginal, j->parent, j->child,
-                                                 lp);
-                            update_sample_list(marginal, j->parent, slp);
-                            ++j;
-                        }
-
-                    double right = maxpos;
-                    if (j < jM)
-                        {
-                            right = std::min(right, j->pos);
-                        }
-                    if (k < kM)
-                        {
-                            right = std::min(right, k->pos);
-                        }
-                    marginal.left = x;
-                    marginal.right = right;
-                    //This "yields"
-                    //the data for this tree
-                    //to the visitor
-                    v(marginal);
-                    x = right;
-                }
-        }
         template <typename visitor>
         void
         algorithmT(const indexed_edge_container& input_left,
